+++ Source/Core/Plugin/NovusPlugin.pas
@@ -2,53 +2,75 @@ unit NovusPlugin;
 
 interface
 
-Uses Windows, SysUtils, Classes, NovusUtilities;
+Uses Windows, SysUtils, Classes, NovusUtilities, NovusList;
 
 const
   func_GetPluginObject = 'GetPluginObject';
 
 type
-  INovusPlugin = interface
-    ['{7AAD37E6-7B50-4266-8268-E8955FC6209E}']
-    function GetPluginName: string; safecall;
+  TNovusPlugin = class(tobject)
+  private
+  protected
+  public
+    function GetPluginName: string; virtual; safecall;
 
-    procedure Initialize; safecall;
-    procedure Finalize; safecall;
+    procedure Initialize; virtual; safecall;
+    procedure Finalize; virtual; safecall;
 
     property PluginName: string read GetPluginName;
   end;
 
-  TGetPluginObject = function: INovusPlugin; stdcall;
+  TGetPluginObject = function: TNovusPlugin; stdcall;
 
-  PPluginInfo = ^TPluginInfo;
+  TPluginInfo = class(TObject)
+  private
+  protected
+    fsFileName: String;
+    fHandle: Thandle;
+    fsPluginName: String;
+    fPlugin: TNovusPlugin;
+    fGetPluginObjectFunc: TGetPluginObject;
+  public
+    destructor Destroy; override;
 
-  TPluginInfo = record
-    FileName: string;
-    PluginName: String;
-    Handle: Thandle;
-    Plugin: INovusPlugin;
-    GetPluginObjectFunc: TGetPluginObject;
+    property FileName: string
+      read fsFilename write fsFilename;
+
+    property PluginName: String
+      read fsPluginName write fsPluginName;
+
+    property Handle: Thandle
+      read fHandle write fHandle;
+
+    property Plugin: TNovusPlugin
+       read fPlugin write fPlugin;
+
+    property GetPluginObjectFunc: TGetPluginObject
+      read fGetPluginObjectFunc write fGetPluginObjectFunc;
   end;
 
   TNovusPlugins = class(Tobject)
   private
   protected
-    fPluginList: TList;
-    function GetPlugins(Index: integer): INovusPlugin;
+    fPluginList: TNovusList;
+    function GetPlugins(Index: integer): TNovusPlugin;
     function GetPluginCount: integer;
   public
     constructor Create;
     destructor Destroy; override;
 
     procedure ClearPluginList;
-    function GetPluginlist(aIndex: Integer): PPluginInfo;
+    function GetPluginlist(aIndex: Integer): tPluginInfo;
 
     function FindPlugin(const aPluginName: string;
-      out aPlugin: INovusPlugin): boolean;
+      out aPlugin: TNovusPlugin): boolean; overload;
+
+    function FindPlugin(const aPluginName: string): boolean; overload;
+
     function LoadPlugin(const aFilename: String): boolean;
     procedure UnloadPlugin(aIndex: integer);
     procedure UnloadAllPlugins;
-    property Plugins[Index: integer]: INovusPlugin read GetPlugins; default;
+    property Plugins[Index: integer]: TNovusPlugin read GetPlugins; default;
     property PluginCount: integer read GetPluginCount;
   end;
 
@@ -58,7 +80,7 @@ uses System.Generics.Defaults;
 
 constructor TNovusPlugins.Create;
 begin
-  fPluginList := TList.Create;
+  fPluginList := TNovusList.Create(TPluginInfo);
 end;
 
 destructor TNovusPlugins.Destroy;
@@ -80,45 +102,31 @@ end;
 
 procedure TNovusPlugins.UnloadPlugin(aIndex: integer);
 var
-  FPluginInfo: PPluginInfo;
+  FPluginInfo: tPluginInfo;
   lHandle: Thandle;
 begin
-  if PluginCount = 0 then
-    Exit;
+  if (aIndex = fPluginList.Count) or (aIndex < 0) then Exit;
 
-  FPluginInfo := fPluginList[aIndex];
-  if (FPluginInfo^.Handle = 0) then
+  FPluginInfo := fPluginList.Items[aIndex] as tPluginInfo;
+  if (FPluginInfo.Handle = 0) then
     Exit;
 
   try
     Try
-      FPluginInfo^.Plugin.Finalize;
-
-      lHandle := FPluginInfo^.Handle;
-
-      FPluginInfo^.Plugin := nil;
-
-      if (lHandle <> 0) then
-        FreeLibrary(lHandle);
+      FPluginInfo.Plugin.Finalize;
     Except
       raise Exception.Create('Unloaded error:' + TNovusUtilities.GetExceptMess);
     End;
 
   finally
-    Dispose(FPluginInfo);
-
-    FPluginInfo := nil;
-
-    fPluginList.Delete(aIndex);
+    fPluginList.Delete(FPluginInfo, true);
   end;
 end;
 
 function TNovusPlugins.LoadPlugin(const aFilename: String): boolean;
 var
   lHandle: Thandle;
-  FPluginInfo: PPluginInfo;
-  ptr: pointer;
-  FPlugin: INovusPlugin;
+  FPluginInfo: tPluginInfo;
 begin
   Result := False;
 
@@ -126,40 +134,34 @@ begin
     Exit;
 
   Try
-    New(FPluginInfo);
+    FPluginInfo := tPluginInfo.Create;
 
-    FPluginInfo^.FileName := aFilename;
+    FPluginInfo.FileName := aFilename;
 
-    FPluginInfo^.Handle := 0;
-    FPluginInfo^.Handle := LoadLibrary(Pchar(aFilename));
+    FPluginInfo.Handle := SafeLoadLibrary(aFilename);
 
-    ptr := GetProcAddress(FPluginInfo^.Handle, func_GetPluginObject);
+    FPluginInfo.GetPluginObjectFunc := GetProcAddress(FPluginInfo.Handle, func_GetPluginObject);
 
-    @FPluginInfo^.GetPluginObjectFunc := ptr;
-    FPlugin := FPluginInfo.GetPluginObjectFunc();
-    FPluginInfo^.Plugin := FPlugin;
-    FPluginInfo^.PluginName := FPlugin.PluginName;
+    FPluginInfo.Plugin := FPluginInfo.GetPluginObjectFunc();
+    FPluginInfo.PluginName := FPluginInfo.Plugin.PluginName;
 
-    if FindPlugin(FPluginInfo^.Plugin.PluginName, FPlugin) then
+    if FindPlugin(FPluginInfo.Plugin.PluginName) then
       raise Exception.Create('Plugin Loaded already');
 
-    FPluginInfo^.Plugin.Initialize;
+    FPluginInfo.Plugin.Initialize;
 
     fPluginList.Add(FPluginInfo);
 
     Result := True;
   Except
-    lHandle := FPluginInfo^.Handle;
-    FPluginInfo^.Plugin := nil;
-    Dispose(FPluginInfo);
-    if (lHandle <> 0) then
-      FreeLibrary(lHandle);
+    if Assigned(FPluginInfo) then FPluginInfo.Free;
+
     raise;
   End;
 end;
 
 function TNovusPlugins.FindPlugin(const aPluginName: string;
-  out aPlugin: INovusPlugin): boolean;
+  out aPlugin: TNovusPlugin): boolean;
 var
   I: integer;
 begin
@@ -170,26 +172,50 @@ begin
     if SameText(Plugins[I].PluginName, aPluginName) then
     begin
       aPlugin := Plugins[I];
-
       Result := True;
+
       Exit;
     end;
 end;
 
 
-function TNovusPlugins.GetPluginlist(aIndex: Integer): PPluginInfo;
+function TNovusPlugins.FindPlugin(const aPluginName: string): boolean;
+var
+  I: integer;
+begin
+  Result := False;
+
+  for I := 0 to (PluginCount - 1) do
+    if SameText(Plugins[I].PluginName, aPluginName) then
+    begin
+      Result := True;
+
+      Exit;
+    end;
+end;
+
+function TNovusPlugins.GetPluginlist(aIndex: Integer): tPluginInfo;
 begin
   Result := NIl;
 
   if (aIndex = fPluginList.Count) or (aIndex < 0) then Exit;
 
-  Result := fPluginList[aIndex];
+  Result := fPluginList.items[aIndex] as tPluginInfo;
 end;
 
-function TNovusPlugins.GetPlugins(Index: integer): INovusPlugin;
+function TNovusPlugins.GetPlugins(Index: integer): TNovusPlugin;
+Var
+  FPluginInfo: tPluginInfo;
 begin
-  Result := PPluginInfo(fPluginList[Index])^.Plugin;
+  Result := NIL;
+
+  if (Index = fPluginList.Count) or (Index < 0) then Exit;
+
+  FPluginInfo := fPluginList.Items[index] as tPluginInfo;
+
+  Result := FPluginInfo.Plugin;
 end;
+
 
 function TNovusPlugins.GetPluginCount: integer;
 begin
@@ -201,4 +227,28 @@ begin
   fPluginList.Clear;
 end;
 
+// TPluginInfo
+destructor TPluginInfo.Destroy;
+begin
+  if Assigned(fPlugin) then fPlugin := NIL;
+
+  if (FHandle <> 0) then FreeLibrary(FHandle);
+end;
+
+
+// TNovusPlugin
+function TNovusPlugin.GetPluginName: string;
+begin
+  result := '';
+end;
+
+procedure TNovusPlugin.Initialize;
+begin
+end;
+
+procedure TNovusPlugin.Finalize;
+begin
+end;
+
 end.
+